requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
aiosqlite==0.20.0
telethon==1.36.0
neo-api-client @ git+https://github.com/Kotak-Neo/kotak-neo-api.git#egg=neo_api_client
pyotp==2.9.0
websockets==12.0
httpx==0.27.0
pandas>=2.0.0
numpy>=1.26.0
